+++ src/main/java/com/msd/erp/application/views/PurchaseOrderDTO.java
@@ -3,7 +3,6 @@ package com.msd.erp.application.views;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -13,6 +12,6 @@ public class PurchaseOrderDTO {
     private Long customerId;
     private Long projectId;
     private LocalDateTime date;
-    private BigDecimal totalPrice;
-    private BigDecimal totalPriceWithVAT;
+    private Double totalPrice;
+    private Double totalPriceWithVAT;
 }
